+++ src/Core/Command/BoredCommand.ts
