+++ src/Core/Command/CommandObject.ts
