+++ src/Core/Command/CryptoCommand.ts
