+++ src/Core/Command/HashCommand.ts
