+++ src/Core/Command/ICallbackable.ts
