+++ src/Core/Command/JokeCommand.ts
