+++ src/Core/Command/LieDetectorCommand.ts
