+++ src/Core/Command/MsiaCovidCommand.ts
