+++ src/Core/Command/PollCommand.ts
