+++ src/Core/Command/TeamCommand.ts
