+++ src/Core/Command/VotingPollCommand.ts
