+++ src/Core/Data/EmojiList.ts
