+++ src/Core/Data/GreetingsData.ts
