+++ src/Core/Data/HdlReplyData.ts
