+++ src/Core/Data/ListEmoji.ts
