+++ src/Core/MessageHandler/ThanksMessageHandler.ts
