+++ src/Core/MessageHandler/VulgarMessageHandler.ts
