+++ src/Core/Schemas/CoinMarketCap.ts
