+++ src/Core/Utility/GeneralUtils.ts
