+++ src/Core/Utility/HttpHeader.ts
